requirements.txt
requests
discord
colorama
subprocess
socket
string
pillow
scapy
pystyle
smtplib
sys
strftime
requests_html
asyncio
aiohttp
gtts
discord.ext